Funciones_JM/requirements.txt
google-generativeai
python-dotenv
Pillow
SpeechRecognition
PyAudio